@ilikerustoo helped with changing imports and renaming of globals to constants
